reloc message and reaction handle to shared
