+++ shared/constants.py
