+++ shared/event_operations.py
@@ -1,6 +1,9 @@
+from emotes.command_operations import send_dm_once
+import discord
+import logging
 import textwrap
 
-async def handle_on_message(bot, message, proposals):
+async def handle_message(bot, message, contributors, emoji_id_mapping, proposals):
     fmt_proposals = ""
 
     # Loop over proposals and convert them to str with every proposal name being on a newline
@@ -10,26 +13,55 @@ async def handle_on_message(bot, message, proposals):
     if message.author == bot.user:
         return
 
+    for emoji_id, contributor_uid in emoji_id_mapping.items():
+        contributor = None
+        for c in contributors:
+            if c["uid"] == contributor_uid:
+                contributor = c
+                break
+
+        if emoji_id in message.content:
+            logging.info('Emoji Found in message! %s', emoji_id)
+
+            if contributor:
+                try:
+                    logging.info(f'Messaging the user, {contributor["uid"]}')
+                    message_link = message.jump_url
+                    await send_dm_once(bot, contributor, message_link)
+                except discord.errors.NotFound:
+                    logging.warning(f'User not found: {contributor["uid"]}')
+
     await bot.process_commands(message)
 
-async def handle_on_reaction_add(bot, user, reaction, proposals, new_proposal_emoji):
+async def handle_reaction(bot, reaction, user, contributors, emoji_id_mapping, proposals, new_proposal_emoji):
+    if user == bot.user:
+        return
+
+    contributor_emoji = next(
+        (emoji_id for emoji_id, contributor_uid in emoji_id_mapping.items() if str(reaction.emoji) == emoji_id),
+        None
+    )
+
+    if contributor_emoji:
+        contributor = next(
+            (c for c in contributors if c["uid"] == emoji_id_mapping[contributor_emoji]),
+            None
+        )
+
+        if contributor:
+            message_link = reaction.message.jump_url
+            logging.info("Emoji react found, DMing contributor")
+            await send_dm_once(bot, contributor, message_link)
 
     def check(m):
         return m.author == user and m.channel == reaction.message.channel
-    
+
     channel = reaction.message.channel
 
     if reaction.emoji == "📝":
-        # Find/Get proposal that the user wants to edit or None
         edit_proposal = next(
-            (
-                item
-                for item in proposals
-                if (
-                    reaction.message.content.strip().endswith(item["name"].strip())
-                )  # In case of conflicts use == comparison after removing emoji and whitespace
-            ),
-            None,
+            (item for item in proposals if reaction.message.content.strip().endswith(item["name"].strip())),
+            None
         )
 
         if edit_proposal:
